[BUGFIX] Fix problem with cache being cleared multiple time
When pages or records in general are changed using
the DataHandler, the pageCache is cleared for
each page multiple times. To fix this, the DataHandler
will collect which pages need clearing, and wait with
the actual clearing until all operations in
cmdMap or dataMap has been processed.
This way the cache is only cleared once for each page.

Resolves: #52235
Releases: 6.2
Change-Id: If1aa389991b801413eafa38d7d140f932bb321f2
Reviewed-on: https://review.typo3.org/25423
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Jigal van Hemert
Reviewed-by: Jan-Erik Revsbech
Tested-by: Jan-Erik Revsbech
Tested-by: Philipp Gampe
Reviewed-by: Benjamin Mack
Tested-by: Benjamin Mack