[BUGFIX] Field size of t3ver_label too low
During workspace move operations sometimes a warning like
this can be found:

  102: These fields are not properly updated in database:
  (t3ver_label) Probably value mismatch with fieldtype.

The reason is, that the t3ver_label field in almost every
record is limited to 30 characters in the SQL definition.
Besides that the error code "102" of the mentioned warning
is invalid and should just be "1" to be recognized by the
accordant functional tests as well.

Resolves: #59110
Releases: 6.2
Change-Id: Ia97a32dbefd1d27e33f186c19c5a6a495dcc194a
Reviewed-on: https://review.typo3.org/30466
Reviewed-by: Wouter Wolters
Reviewed-by: Krzysztof Adamczyk
Reviewed-by: Marcin Sągol
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader