[BUGFIX] Correct dependency handling in workspaces module
This patch fixes a bug where records, which have references
pointing to them, are in a table that is disabled for
workspaces throw an element not found RuntimeException.

Without this change, you will get an error in the workspace module
for tables that are not workspace-enabled, but have file references
pointing to them.

Change-Id: I16c42f0347e0c1c6d916e24f865eb2af643929da
Resolves: #66523
Releases: master, 6.2
Reviewed-on: http://review.typo3.org/38858
Reviewed-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Tested-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>