[!!!][TASK] Remove various TSFE properties
The short-hand syntax for TYPO3_CONF_VARS in TSFE
is rather tedious when explaining to newcomers why
it was solved this way 12 years ago.

The patch changes all occurrences to use the
global array, as in all other parts of the TYPO3 Core.

Additionally, the following public properties are removed:

* defaultBodyTag (can be overriden via TypoScript)
* clientInfo (can be loaded via GeneralUtility::clientInfo())

Resolves: #77184
Releases: master
Change-Id: I7d6a78f0d32db5fb7635bb7dc7770cae49925644
Reviewed-on: https://review.typo3.org/49135
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Bamboo TYPO3com <info@typo3.com>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>