[TASK] Streamline assignment of current workspace in DataHandler
The processing of data structures in DataHandler and RelationHandler
are bound to each other. DataHandler allows to be initialized with a
custom BE_USER instance and thus with a custom workspace, different
to the actual value of the real instance of $GLOBALS['BE_USER'].

This patch streamlines the workspace disposal by explicitly
forwarding the provided workspace to RelationHandler and avoids
fetching the current workspace for processing from global values.

The consideration of workspace related fields in RelationHandler was
bound to the current workspace of the global backend user - however,
this backend user constraint has been removed now.

Resolves: #79339
Releases: master
Change-Id: I3973d55c5b9d99ed2e976b86136338fee79d5f6e
Reviewed-on: https://review.typo3.org/51325
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Susanne Moog <susanne.moog@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Susanne Moog <susanne.moog@typo3.org>